[KHJ] LinkRegisterReqDto 내 자료형 변경 (int → long)

+++ ptis-core/src/main/java/com/thewayhome/ptis/core/dto/LinkRegisterReqDto.java
@@ -15,5 +15,5 @@ public class LinkRegisterReqDto extends BaseRequestDto {
     private String linkName;
     private Node stNode;
     private Node edNode;
-    private int cost;
+    private long cost;
 }
